+++ src/main/java/com/handwoong/rainbowletter/domain/faq/dto/FAQRequest.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.faq;
+package com.handwoong.rainbowletter.domain.faq.dto;
 
 import static com.handwoong.rainbowletter.dto.ValidateMessage.EMPTY_MESSAGE;
 import static com.handwoong.rainbowletter.dto.ValidateMessage.FAQ_SUMMARY;
