+++ src/main/java/com/handwoong/rainbowletter/domain/faq/dto/FAQResponse.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.dto.faq;
+package com.handwoong.rainbowletter.domain.faq.dto;
 
 import java.util.List;
