+++ src/main/java/com/handwoong/rainbowletter/domain/faq/repository/FAQRepository.java
@@ -1,8 +1,8 @@
-package com.handwoong.rainbowletter.repository.faq;
+package com.handwoong.rainbowletter.domain.faq.repository;
 
-import com.handwoong.rainbowletter.domain.faq.FAQ;
-import com.handwoong.rainbowletter.dto.faq.FAQAdminDto;
-import com.handwoong.rainbowletter.dto.faq.FAQDto;
+import com.handwoong.rainbowletter.domain.faq.dto.FAQAdminDto;
+import com.handwoong.rainbowletter.domain.faq.dto.FAQDto;
+import com.handwoong.rainbowletter.domain.faq.model.FAQ;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
